runpod/requirements.txt
# RunPod Serverless Dependencies
runpod>=1.6.0
openai-whisper>=20231117
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
sentencepiece>=0.1.99
accelerate>=0.24.0
